attached update.report_out to DirectImageGuide
