+++ src/pytti/workhorse.py
@@ -508,59 +508,6 @@ def _main(cfg: DictConfig):
             params=params,
             writer=writer,
         ):
-            def report_out(
-                i,
-                stage_i,
-                model,
-                writer,
-                fig,  # default to None...
-                axs,  # default to None...
-                clear_every,
-                display_every,
-                approximate_vram_usage,
-                display_scale,
-                show_graphs,
-                show_palette,
-            ):
-
-                # DM: I bet this could be abstracted out into a report_out() function or whatever
-                if clear_every > 0 and i > 0 and i % clear_every == 0:
-                    display.clear_output()
-
-                if display_every > 0 and i % display_every == 0:
-                    logger.debug(f"Step {i} losses:")
-                    if model.dataframe:
-                        rec = model.dataframe[0].iloc[-1]
-                        logger.debug(rec)
-                        for k, v in rec.iteritems():
-                            writer.add_scalar(
-                                tag=f"losses/{k}", scalar_value=v, global_step=i
-                            )
-
-                    # does this VRAM stuff even do anything?
-                    if approximate_vram_usage:
-                        logger.debug("VRAM Usage:")
-                        print_vram_usage()  # update this function to use logger
-                    # update this stuff to use/rely on tensorboard
-                    display_width = int(img.image_shape[0] * display_scale)
-                    display_height = int(img.image_shape[1] * display_scale)
-                    if stage_i > 0 and show_graphs:
-                        model.plot_losses(axs)
-                        im = img.decode_image()
-                        sidebyside = make_hbox(
-                            im.resize((display_width, display_height), Image.LANCZOS),
-                            fig,
-                        )
-                        display.display(sidebyside)
-                    else:
-                        im = img.decode_image()
-                        display.display(
-                            im.resize((display_width, display_height), Image.LANCZOS)
-                        )
-                    if show_palette and isinstance(img, PixelImage):
-                        logger.debug("Palette:")
-                        display.display(img.render_pallet())
-
             def save_out(
                 i,
                 img,
@@ -610,10 +557,10 @@ def _main(cfg: DictConfig):
 
             ###########################################################
 
-            report_out(
+            model.report_out(
                 i=i,
                 stage_i=stage_i,
-                model=model,
+                # model=model,
                 writer=writer,
                 fig=fig,  # default to None...
                 axs=axs,  # default to None...
